index: Move index DBs into index/ directory.
Summary:
backports last two commits from PR13243:
ec3073a27 index: Move index DBs into index/ directory.
89eddcd36 index: Remove TxIndexDB from public interface of TxIndex.
https://github.com/bitcoin/bitcoin/compare/89eddcd36~1..ec3073a27

(these diffs are co-dependent and need to be brought in as one)

Depends on D2933 D2934 D2937 , and D2932

Test Plan: `make check`

Reviewers: #bitcoin_abc, deadalnix, Fabien, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T630

Differential Revision: https://reviews.bitcoinabc.org/D2938
