+++ src/index/base.h
@@ -5,10 +5,10 @@
 #ifndef BITCOIN_INDEX_BASE_H
 #define BITCOIN_INDEX_BASE_H
 
+#include <dbwrapper.h>
 #include <primitives/block.h>
 #include <primitives/transaction.h>
 #include <threadinterrupt.h>
-#include <txdb.h>
 #include <uint256.h>
 #include <validationinterface.h>
 
@@ -20,6 +20,19 @@ class CBlockIndex;
  * to their position in the active chain.
  */
 class BaseIndex : public CValidationInterface {
+protected:
+    class DB : public CDBWrapper {
+    public:
+        DB(const fs::path &path, size_t n_cache_size, bool f_memory = false,
+           bool f_wipe = false, bool f_obfuscate = false);
+
+        /// Read block locator of the chain that the txindex is in sync with.
+        bool ReadBestBlock(CBlockLocator &locator) const;
+
+        /// Write block locator of the chain that the txindex is in sync with.
+        bool WriteBestBlock(const CBlockLocator &locator);
+    };
+
 private:
     /// Whether the index is in sync with the main chain. The flag is flipped
     /// from false to true once, after which point this starts processing
@@ -58,7 +71,7 @@ protected:
         return true;
     }
 
-    virtual BaseIndexDB &GetDB() const = 0;
+    virtual DB &GetDB() const = 0;
 
     /// Get the name of the index for display in logs.
     virtual const char *GetName() const = 0;
