+++ src/index/txindex.h
@@ -6,6 +6,7 @@
 #define BITCOIN_INDEX_TXINDEX_H
 
 #include <index/base.h>
+#include <txdb.h>
 
 /**
  * TxIndex is used to look up transactions included in the blockchain by hash.
@@ -13,8 +14,11 @@
  * location of each transaction by transaction hash.
  */
 class TxIndex final : public BaseIndex {
+protected:
+    class DB;
+
 private:
-    const std::unique_ptr<TxIndexDB> m_db;
+    const std::unique_ptr<DB> m_db;
 
 protected:
     /// Override base class init to migrate from old database.
@@ -22,13 +26,18 @@ protected:
 
     bool WriteBlock(const CBlock &block, const CBlockIndex *pindex) override;
 
-    BaseIndexDB &GetDB() const override;
+    BaseIndex::DB &GetDB() const override;
 
     const char *GetName() const override { return "txindex"; }
 
 public:
     /// Constructs the index, which becomes available to be queried.
-    explicit TxIndex(std::unique_ptr<TxIndexDB> db);
+    explicit TxIndex(size_t n_cache_size, bool f_memory = false,
+                     bool f_wipe = false);
+
+    // Destructor is declared because this class contains a unique_ptr to an
+    // incomplete type.
+    virtual ~TxIndex() override;
 
     /// Look up a transaction by hash.
     ///
