+++ src/test/txindex_tests.cpp
@@ -14,7 +14,7 @@
 BOOST_AUTO_TEST_SUITE(txindex_tests)
 
 BOOST_FIXTURE_TEST_CASE(txindex_initial_sync, TestChain100Setup) {
-    TxIndex txindex(std::make_unique<TxIndexDB>(1 << 20, true));
+    TxIndex txindex(1 << 20, true);
 
     CTransactionRef tx_disk;
     uint256 block_hash;
